+++ src/components/CurrentUserMessage.tsx
@@ -1,34 +1,12 @@
 import { styled } from '@linaria/react';
-import { SendingStatus, UserMessage } from '@sendbird/chat/message';
+import { UserMessage } from '@sendbird/chat/message';
+import { useTheme } from 'styled-components';
 
-import { DefaultSentTime, BodyContainer, BodyComponent } from './MessageComponent';
+import { BodyContainer, BodyComponent } from './MessageComponent';
+import MyMessageStatus from './MyMessageStatus';
 import { useConstantState } from '../context/ConstantContext';
-import { useWidgetSetting } from '../context/WidgetSettingContext';
-import { Icon } from '../foundation/components/Icon';
-import { formatCreatedAtToAMPM } from '../utils/messageTimestamp';
 
-const SpinnerSize = '16px';
-
-interface IconContainerProps {
-  isSpinning?: boolean;
-}
-
-const IconContainer = styled.div<IconContainerProps>`
-  display: grid;
-  justify-content: center;
-  align-items: center;
-  margin-bottom: 2px;
-  animation: ${({ isSpinning }) => (isSpinning ? 'spinner 1.5s linear infinite' : 'unset')};
-
-  @keyframes spinner {
-    from {
-      transform: rotate(0deg);
-    }
-    to {
-      transform: rotate(360deg);
-    }
-  }
-`;
+import '../css/index.css';
 
 const Root = styled.div<{ enableEmojiFeedback: boolean }>`
   display: flex;
@@ -43,31 +21,13 @@ type Props = {
 
 export default function CurrentUserMessage(props: Props) {
   const { enableEmojiFeedback, dateLocale } = useConstantState();
-  const { theme } = useWidgetSetting();
-  const { message } = props;
+  const theme = useTheme();
 
-  const MessageStatus = () => {
-    switch (message.sendingStatus) {
-      case SendingStatus.PENDING:
-        return (
-          <IconContainer isSpinning={true}>
-            <Icon type={"spinner"} color={theme.bgColor.outgoingMessage} size={16} />
-          </IconContainer>
-        );
-      case SendingStatus.FAILED:
-        return (
-          <IconContainer>
-            <Icon type={"error"} color={"error"} size={16} />
-          </IconContainer>
-        );
-      default:
-        return <DefaultSentTime>{formatCreatedAtToAMPM(message.createdAt, dateLocale)}</DefaultSentTime>;
-    }
-  };
+  const { message } = props;
 
   return (
     <Root enableEmojiFeedback={enableEmojiFeedback}>
-      <MessageStatus />
+      <MyMessageStatus message={message} dateLocale={dateLocale} theme={theme} />
       <BodyContainer>
         <BodyComponent>
           <div className="sendbird-word">{message.message}</div>
