+++ src/components/widget/ProviderContainer.tsx
@@ -5,10 +5,12 @@ import { StyleSheetManager, ThemeProvider } from 'styled-components';
 import SendbirdProvider from '@uikit/lib/Sendbird';
 
 import { ChatAiWidgetProps } from './ChatAiWidget';
+import { generateCSSVariables } from '../../colors';
 import { ConstantStateProvider, useConstantState } from '../../context/ConstantContext';
 import { useWidgetSession, useWidgetSetting, WidgetSettingProvider } from '../../context/WidgetSettingContext';
 import { useWidgetState, WidgetStateProvider } from '../../context/WidgetStateContext';
 import { useStyledComponentsTarget } from '../../hooks/useStyledComponentsTarget';
+import { getTheme } from '../../theme';
 import { DragDropProvider } from '../../tools/hooks/useDragDropFiles';
 import { isDashboardPreview } from '../../utils';
 
@@ -32,7 +34,7 @@ const SBComponent = ({ children }: { children: React.ReactElement }) => {
   } = useConstantState();
 
   const { setIsVisible } = useWidgetState();
-  const { botConfigs, botStyle, theme, colorSet } = useWidgetSetting();
+  const { botConfigs, botStyle } = useWidgetSetting();
   const session = useWidgetSession();
   const target = useStyledComponentsTarget();
 
@@ -49,9 +51,27 @@ const SBComponent = ({ children }: { children: React.ReactElement }) => {
     return userAgent;
   }, []);
 
+  const { theme, accentColor, botMessageBGColor } = botStyle;
+
+  const styledTheme = getTheme({
+    accentColor,
+    botMessageBGColor,
+  })[theme];
+
+  const customColorSet = useMemo(() => {
+    if (!accentColor) return undefined;
+
+    return ['light', 'dark'].reduce((acc, cur) => {
+      return {
+        ...acc,
+        ...generateCSSVariables(accentColor, cur),
+      };
+    }, {});
+  }, [accentColor]);
+
   return (
     <StyleSheetManager target={target}>
-      <ThemeProvider theme={theme}>
+      <ThemeProvider theme={styledTheme}>
         {applicationId && session.userId && (
           <SendbirdProvider
             appId={applicationId}
@@ -64,7 +84,7 @@ const SBComponent = ({ children }: { children: React.ReactElement }) => {
             customExtensionParams={userAgentCustomParams}
             breakpoint={isMobileView} // A property that determines whether to show it with a layout that fits the mobile screen. Or you can put the width size with `px`.
             theme={botStyle.theme}
-            colorSet={colorSet}
+            colorSet={customColorSet}
             stringSet={stringSet}
             dateLocale={dateLocale}
             eventHandlers={{
